refactored EntityTypes and brought back JSON for Entities
