+++ test/Poseidon/EntitiesListSpec.hs
@@ -2,20 +2,17 @@
 
 module Poseidon.EntitiesListSpec (spec) where
 
-import           Poseidon.EntitiesList
-import           Poseidon.EntityTypes  (EntitiesList, HasNameAndVersion (..),
-                                        IndividualInfo (..),
-                                        PacNameAndVersion (..),
-                                        PoseidonEntity (..), SignedEntitiesList,
-                                        SignedEntity (..))
-import           Poseidon.Package      (PackageReadOptions (..),
-                                        defaultPackageReadOptions,
-                                        getJointIndividualInfo,
-                                        readPoseidonPackageCollection)
-import           Poseidon.Utils        (PoseidonException, testLog)
+import           Poseidon.EntityTypes
+import           Poseidon.Package     (PackageReadOptions (..),
+                                       defaultPackageReadOptions,
+                                       filterToRelevantPackages,
+                                       getJointIndividualInfo,
+                                       readPoseidonPackageCollection)
+import           Poseidon.Utils       (PoseidonException, testLog)
 
-import           Data.Either           (fromRight, isLeft)
-import           Data.Version          (makeVersion)
+import           Data.Aeson           (decode, encode)
+import           Data.Either          (fromRight, isLeft)
+import           Data.Version         (makeVersion)
 import           Test.Hspec
 
 spec :: Spec
@@ -26,6 +23,8 @@ spec = do
     testFilterPackages
     testResolveEntityIndices
     testShow
+    testJSON
+
 
 testReadPoseidonEntitiesString :: Spec
 testReadPoseidonEntitiesString =
@@ -346,3 +345,41 @@ testShow =
                     "-Group1"
             show (Exclude (Pac (PacNameAndVersion "Pac1" Nothing))) `shouldBe`
                     "-*Pac1*"
+
+testJSON :: Spec
+testJSON =
+    describe "Poseidon.EntitiesList.ToJSON" $ do
+        it "should encode entities correctly to JSON" $ do
+            encode (Ind "Ind1") `shouldBe`
+                    "\"<Ind1>\""
+            encode (SpecificInd "a" "b" (PacNameAndVersion "c" Nothing)) `shouldBe`
+                    "\"<c:b:a>\""
+            encode (Group "Group1") `shouldBe`
+                    "\"Group1\""
+            encode (Pac (PacNameAndVersion "Pac1" Nothing)) `shouldBe`
+                    "\"*Pac1*\""
+            encode (Exclude (Ind "Ind1")) `shouldBe`
+                    "\"-<Ind1>\""
+            encode (Exclude (SpecificInd "a" "b" (PacNameAndVersion "c" Nothing))) `shouldBe`
+                    "\"-<c:b:a>\""
+            encode (Exclude (Group "Group1")) `shouldBe`
+                    "\"-Group1\""
+            encode (Exclude (Pac (PacNameAndVersion "Pac1" Nothing))) `shouldBe`
+                    "\"-*Pac1*\""
+        it "should decode entities correctly from JSON" $ do
+            decode "\"<Ind1>\""   `shouldBe`
+                    Just (Ind "Ind1")
+            decode "\"<c:b:a>\""  `shouldBe`
+                    Just (SpecificInd "a" "b" (PacNameAndVersion "c" Nothing))
+            decode "\"Group1\""   `shouldBe`
+                    Just (Group "Group1")
+            decode "\"*Pac1*\""   `shouldBe`
+                    Just (Pac (PacNameAndVersion "Pac1" Nothing))
+            decode "\"-<Ind1>\""  `shouldBe`
+                    Just (Exclude (Ind "Ind1"))
+            decode "\"-<c:b:a>\"" `shouldBe`
+                    Just (Exclude (SpecificInd "a" "b" (PacNameAndVersion "c" Nothing)))
+            decode "\"-Group1\""  `shouldBe`
+                    Just (Exclude (Group "Group1"))
+            decode "\"-*Pac1*\""  `shouldBe`
+                    Just (Exclude (Pac (PacNameAndVersion "Pac1" Nothing)))
